update taxonomy tests
